backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
httpx>=0.25.0
python-multipart>=0.0.6
aiofiles>=23.2.0
chromadb>=0.4.15
sentence-transformers>=2.2.0
numpy>=1.21.0
pandas>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0
